feat(api): use create assessment result line
